client/src/app/services/chat-service/chat.service.ts
import { Injectable } from '@angular/core';
import { SocketsOnEvents, SocketsSendEvents } from '@app/constants/constants';
import { Message } from '@app/interfaces/message';
import { SocketService } from '@app/services/socket-service/socket.service';

/**
 * This class allows to handle the business logic for the chat through all the components
 * where he is used.
 */
@Injectable({
    providedIn: 'root',
})
export class ChatService {
    messagesList: Message[] = [];
    hasJustSentMessage: boolean = false;
    isChatAccessible: boolean = true;

    constructor(private socketService: SocketService) {}

    send(msg: Message): void {
        if (this.isChatAccessible) this.socketService.send<Message>(SocketsSendEvents.SendMessage, msg);
    }

    setupListeners(): void {
        this.socketService.on<Message>(SocketsOnEvents.ChatMessage, (msg) => {
            this.messagesList.push(msg);
            this.hasJustSentMessage = true;
        });
    }

    cleanMessages(): void {
        this.messagesList = [];
    }
}
